rename: 불필요한 코드 제거 & 네이밍 수정
