+++ app/src/main/java/woowacourse/shopping/di/DaoBinder.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.room.Room
 import com.woowa.di.ApplicationContext
 import com.woowa.di.component.InstallIn
-import com.woowa.di.singleton.SingletonComponent2
+import com.woowa.di.singleton.SingletonComponent
 import woowacourse.shopping.data.CartProductDao
 import woowacourse.shopping.data.ShoppingDatabase
 import javax.inject.Qualifier
@@ -15,7 +15,7 @@ annotation class InMemory
 @Qualifier
 annotation class Database
 
-@InstallIn(SingletonComponent2::class)
+@InstallIn(SingletonComponent::class)
 class DaoBinder {
     @Database
     fun provideCartProductDao(
