+++ di/src/main/java/com/woowa/di/component/Component.kt
@@ -1,10 +1,21 @@
 package com.woowa.di.component
 
 import kotlin.reflect.KClass
+import kotlin.reflect.KFunction
 
 interface Component {
     fun getDIInstance(
         type: KClass<*>,
         qualifier: KClass<out Annotation>? = null,
     ): Any?
+
+    fun registerDIInstance(
+        binder: Any,
+        kFunc: KFunction<*>,
+    )
+
+    fun deleteDIInstance(
+        type: KClass<*>,
+        qualifier: KClass<out Annotation>?,
+    )
 }
