+++ di/src/main/java/com/woowa/di/component/InstallIn.kt
@@ -4,4 +4,4 @@ import kotlin.reflect.KClass
 
 @Target(AnnotationTarget.CLASS)
 @Retention(AnnotationRetention.RUNTIME)
-annotation class InstallIn(val component2: KClass<out Component2>)
+annotation class InstallIn(val component: KClass<out Component>)
