+++ di/src/main/java/com/woowa/di/component/ParentManager.kt
@@ -1,0 +1,15 @@
+package com.woowa.di.component
+
+import kotlin.reflect.KClass
+
+
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class ParentManager2(val manager: KClass<out ComponentManager>)
+
+
+data object NoParent : ComponentManager() {
+    override fun <T : Any> getComponentInstance(componentType: KClass<out T>): Component {
+        throw IllegalArgumentException()
+    }
+}
