+++ di/src/main/java/com/woowa/di/singleton/SingletonComponent.kt
@@ -1,5 +1,6 @@
 package com.woowa.di.singleton
 
+import android.app.Application
 import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.ProcessLifecycleOwner
@@ -10,64 +11,71 @@ import com.woowa.di.findQualifierClassOrNull
 import javax.inject.Inject
 import kotlin.reflect.KClass
 import kotlin.reflect.KFunction
-import kotlin.reflect.KVisibility
-import kotlin.reflect.full.createInstance
-import kotlin.reflect.full.declaredMemberFunctions
+import kotlin.reflect.KParameter
 import kotlin.reflect.full.hasAnnotation
 import kotlin.reflect.jvm.jvmErasure
 import kotlin.reflect.jvm.kotlinProperty
 
-class SingletonComponent<binder : Any> private constructor(private val binderClazz: KClass<binder>) :
+class SingletonComponent<T : Application> private constructor(private val clazz: KClass<T>) :
     Component, DefaultLifecycleObserver {
-        private lateinit var binderInstance: binder
-        private lateinit var diInstances: Map<String, Any?>
-
-        override fun onCreate(owner: LifecycleOwner) {
-            super.onCreate(owner)
-            binderInstance = binderClazz.createInstance()
-            diInstances = saveDIInstance()
-        }
+        private val diFunc: MutableMap<KFunction<*>, Any> = mutableMapOf()
+        private val diInstances: MutableMap<String, Any?> = mutableMapOf()
 
         override fun onDestroy(owner: LifecycleOwner) {
             super.onDestroy(owner)
-            instances.remove(binderClazz)
-            owner.lifecycle.removeObserver(this)
-        }
-
-        private fun saveDIInstance(): Map<String, Any?> {
-            return binderClazz.declaredMemberFunctions.filter { it.visibility == KVisibility.PUBLIC }
-                .associate { kFunc ->
-                    val key =
-                        kFunc.returnType.jvmErasure.simpleName
-                            ?: error("익명 객체와 같이, 이름이 없는 객체는 di 주입을 할 수 없습니다.")
-
-                    val result = createInjectedInstance(kFunc)
-
-                    kFunc.findQualifierClassOrNull()?.let { qualifier ->
-                        return@associate (key + qualifier.simpleName) to result
-                    }
-
-                    return@associate key to result
-                }
-        }
-
-        private fun createInjectedInstance(kFunc: KFunction<*>): Any? {
-            return if (kFunc.parameters.any { it.hasAnnotation<ApplicationContext>() }) {
-                kFunc.call(binderInstance, DIBuilder.applicationContext)
-            } else {
-                kFunc.call(binderInstance)
-            }
+            diInstances.clear()
+            instances.remove(clazz)
+            ProcessLifecycleOwner.get().lifecycle.removeObserver(this)
         }
 
         override fun getDIInstance(
             type: KClass<*>,
             qualifier: KClass<out Annotation>?,
         ): Any? {
+            qualifier?.let {
+                return diInstances.getOrPut((type.simpleName + it.simpleName)) {
+                    createDIInstance(type, qualifier)
+                }
+            }
+            return diInstances.getOrPut(
+                type.simpleName ?: error("익명 객체와 같이, 이름이 없는 객체는 di 주입을 할 수 없습니다."),
+            ) {
+                createDIInstance(type)
+            }
+        }
+
+        override fun registerDIInstance(
+            binder: Any,
+            kFunc: KFunction<*>,
+        ) {
+            diFunc[kFunc] = binder
+        }
+
+        private fun createDIInstance(
+            type: KClass<*>,
+            qualifier: KClass<out Annotation>? = null,
+        ): Any? {
+            val kFunc =
+                diFunc.keys.find { it.returnType.jvmErasure == type && it.findQualifierClassOrNull() == qualifier }
+                    ?: error("${type.simpleName}이 component에 등록되지 않았습니다.")
+
+            val parameters =
+                kFunc.parameters.filter { it.kind == KParameter.Kind.VALUE }.map {
+                    when {
+                        it.hasAnnotation<ApplicationContext>() -> DIBuilder.applicationContext
+                        else ->
+                            SingletonComponentManager.getDIInstance(
+                                it.type.jvmErasure,
+                                it.findQualifierClassOrNull(),
+                            )
+                    }
+                }
+
             val instance =
-                if (qualifier != null) {
-                    diInstances[(type.simpleName + qualifier.simpleName)]
+                if (parameters.isEmpty()) {
+                    kFunc.call(diFunc[kFunc])
                 } else {
-                    diInstances[type.simpleName]
+                    kFunc.call(diFunc[kFunc], *parameters.toTypedArray())
                 }
 
             injectFields(instance)
@@ -90,12 +98,24 @@ class SingletonComponent<binder : Any> private constructor(private val binderCla
             }
         }
 
+        override fun deleteDIInstance(
+            type: KClass<*>,
+            qualifier: KClass<out Annotation>?,
+        ) {
+            qualifier?.let {
+                diInstances.remove(type.simpleName + qualifier.simpleName)
+                return
+            }
+            diInstances.remove(type.simpleName)
+        }
+
         companion object {
             private val instances = mutableMapOf<KClass<*>, SingletonComponent<*>>()
 
-            fun <binder : Any> getInstance(binderClazz: KClass<binder>): SingletonComponent<binder> {
+            fun <binder : Application> getInstance(binderClazz: KClass<binder>): SingletonComponent<binder> {
                 return instances.getOrPut(binderClazz) {
                     val newInstance = SingletonComponent(binderClazz)
+                    instances[binderClazz] = newInstance
                     ProcessLifecycleOwner.get().lifecycle.addObserver(newInstance)
                     newInstance
                 } as SingletonComponent<binder>
