+++ di/src/main/java/com/woowa/di/test/DIActivityTestRule.kt
@@ -6,9 +6,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.ViewModelStore
 import androidx.lifecycle.testing.TestLifecycleOwner
 import com.woowa.di.singleton.SingletonComponent
-import com.woowa.di.singleton.SingletonComponent2
 import com.woowa.di.singleton.SingletonComponentManager
-import com.woowa.di.singleton.SingletonComponentManager2
 import org.junit.rules.ExternalResource
 import org.robolectric.Robolectric
 import org.robolectric.android.controller.ActivityController
@@ -25,8 +23,8 @@ class DIActivityTestRule<T : ComponentActivity>(private val activityClass: Class
         super.before()
         applicationLifecycleOwner = TestLifecycleOwner(Lifecycle.State.CREATED)
 
-        SingletonComponentManager2.binderClazzs.forEach {
-            val component = SingletonComponent2.getInstance(it as KClass<Application>)
+        SingletonComponentManager.binderClazzs.forEach {
+            val component = SingletonComponent.getInstance(it as KClass<Application>)
             applicationLifecycleOwner.lifecycle.addObserver(component)
         }
         applicationLifecycleOwner.handleLifecycleEvent(Lifecycle.Event.ON_CREATE)
