+++ di/src/main/java/com/woowa/di/viewmodel/DIViewModelFactory.kt
@@ -7,7 +7,7 @@ import kotlin.reflect.full.createInstance
 
 inline fun <reified T : ViewModel> getDIViewModelFactory(): ViewModelProvider.Factory {
     val instance = T::class.createInstance()
-    ViewModelComponentManager2.createComponent(T::class)
+    ViewModelComponentManager.createComponent(T::class)
     injectViewModelComponentFields<T>(instance)
     removeInstancesOnCleared<T>(instance)
     return viewModelFactory {
