+++ di/src/test/java/com/woowa/di/fixture/component/SingletonComponentTestBinder.kt
@@ -1,9 +1,9 @@
 package com.woowa.di.fixture.component
 
 import com.woowa.di.component.InstallIn
-import com.woowa.di.singleton.SingletonComponent2
+import com.woowa.di.singleton.SingletonComponent
 
-@InstallIn(SingletonComponent2::class)
+@InstallIn(SingletonComponent::class)
 class SingletonComponentTestBinder {
     fun provideSingletonComponent(): TestSingletonComponent = FakeSingleComponentImpl()
 }
