+++ di/src/test/java/com/woowa/di/fixture/component/ViewModelComponentTestBinder.kt
@@ -1,9 +1,9 @@
 package com.woowa.di.fixture.component
 
 import com.woowa.di.component.InstallIn
-import com.woowa.di.viewmodel.ViewModelComponent2
+import com.woowa.di.viewmodel.ViewModelComponent
 
-@InstallIn(ViewModelComponent2::class)
+@InstallIn(ViewModelComponent::class)
 class ViewModelComponentTestBinder {
     fun provideViewModelComponent(): TestViewModelComponent = FakeViewModelComponentImpl()
 }
